MVP-4153: Wallet custom hook/ Major refactor
