+++ packages/notifi-dapp-example/src/app/layout.tsx
@@ -1,31 +1,24 @@
-'use client';
-
 import { GlobalStateContextProvider } from '@/context/GlobalStateContext';
 import { CosmosWalletProvider } from '@/context/wallet/CosmosWalletProvider';
-// import NotifiWalletProvider from '@/context/wallet/NotifiWalletProvider';
+import { NotifiWalletsWrapper } from '@/context/wallet/NotifiWalletsWrapper';
 import '@interchain-ui/react/styles';
-import {
-  NotifiWalletProvider,
-  NotifiWalletsContext,
-} from '@notifi-network/notifi-wallet-provider';
-// import type { Metadata } from 'next';
+import type { Metadata } from 'next';
 import { Inter } from 'next/font/google';
 
 import './globals.css';
 
 const inter = Inter({ subsets: ['latin'] });
 
-// const metadata: Metadata = {
-//   title: 'Notifi Dapp Example',
-//   description: 'Notifi Dapp Example',
-// };
+export const metadata: Metadata = {
+  title: 'Notifi Dapp Example',
+  description: 'Notifi Dapp Example',
+};
 
 export default function RootLayout({
   children,
 }: Readonly<{
   children: React.ReactNode;
 }>) {
-  // console.log('TT', TT);
   return (
     <html lang="en">
       <head>
@@ -37,11 +30,9 @@ export default function RootLayout({
         // NOTE: Not support dark/light mode yet. TODO: consider adding a theme wrapper
       >
         <GlobalStateContextProvider>
-          <NotifiWalletProvider>
-            {/* <NotifiWalletsContext> */}
+          <NotifiWalletsWrapper>
             <CosmosWalletProvider>{children}</CosmosWalletProvider>
-            {/* </NotifiWalletsContext> */}
-          </NotifiWalletProvider>
+          </NotifiWalletsWrapper>
         </GlobalStateContextProvider>
       </body>
     </html>
