+++ packages/notifi-wallet-provider/lib/index.ts
@@ -1,1 +1,2 @@
 export * from './context';
+export * as Types from './types';
